feat : implemented course cubit and created it's tests
