--- lib/core/common/app/features/course/domain/repositories/course_repo.dart
@@ -1,9 +1,0 @@
-import 'package:education_app/core/common/app/features/course/domain/entities/course.dart';
-import 'package:education_app/core/utils/typedefs.dart';
-
-abstract class CourseRepo {
-  const CourseRepo();
-
-  ResultFuture<List<Course>> getCourses();
-  ResultFuture<void> addCourse(Course course);
-}
