--- lib/core/common/app/features/course/domain/usecases/add_course.dart
@@ -1,15 +1,0 @@
-import 'package:education_app/core/common/app/features/course/domain/entities/course.dart';
-import 'package:education_app/core/common/app/features/course/domain/repositories/course_repo.dart';
-import 'package:education_app/core/usecases/usecases.dart';
-import 'package:education_app/core/utils/typedefs.dart';
-
-class AddCourse extends UsecaseWithParams<void, Course> {
-  const AddCourse(this._repo);
-
-  final CourseRepo _repo;
-
-  @override
-  ResultFuture<void> call(Course params) {
-    return _repo.addCourse(params);
-  }
-}
