+++ lib/src/course/domain/entities/course.dart
