+++ lib/src/course/domain/repositories/course_repo.dart
@@ -1,0 +1,9 @@
+import 'package:education_app/src/course/domain/entities/course.dart';
+import 'package:education_app/core/utils/typedefs.dart';
+
+abstract class CourseRepo {
+  const CourseRepo();
+
+  ResultFuture<List<Course>> getCourses();
+  ResultFuture<void> addCourse(Course course);
+}
