--- test/core/common/app/features/course/domain/usecases/course_usecases_mock.dart
@@ -1,4 +1,0 @@
-import 'package:education_app/core/common/app/features/course/domain/repositories/course_repo.dart';
-import 'package:mocktail/mocktail.dart';
-
-class CourseRepoMock extends Mock implements CourseRepo {}
