+++ test/src/course/domain/usecases/course_usecases_mock.dart
@@ -1,0 +1,4 @@
+import 'package:education_app/src/course/domain/repositories/course_repo.dart';
+import 'package:mocktail/mocktail.dart';
+
+class CourseRepoMock extends Mock implements CourseRepo {}
